fetch_catalog/requirements.txt
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
flake8==3.8.4
GeoAlchemy2==0.8.4
idna==2.10
importlib-metadata==3.3.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.9
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.3
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
SQLAlchemy==1.3.22
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
